Checkpoint version

+++ EasyBank/BAL/HistoryManager.cs
@@ -9,17 +9,17 @@ namespace EasyBank.DAL
 {
     public class HistoryManager //Has no db.SaveChanges()!!
     {
-        public static void AddDepositOperation(ConnectionContext db, int amount, int toId, int operatorId)
+        public static void AddDepositOperation(ConnectionContext db, decimal amount, int toId, int operatorId)
         {
             db.OperationHistory.Add(new Operation(amount, OperationTypes.Deposit, toAccountId:toId, operatorId:operatorId));
         }
 
-        public static void AddWidthdrawOperation(ConnectionContext db, int amount, int fromId, int operatorId)
+        public static void AddWidthdrawOperation(ConnectionContext db, decimal amount, int fromId, int operatorId)
         {
             db.OperationHistory.Add(new Operation(-amount, OperationTypes.Withdraw, fromAccountId: fromId, operatorId: operatorId));
         }
 
-        public static void AddTransferOperation(ConnectionContext db, int amount, int fromId, int toId, int operatorId)
+        public static void AddTransferOperation(ConnectionContext db, decimal amount, int fromId, int toId, int operatorId)
         {
             db.OperationHistory.Add(new Operation(amount, OperationTypes.Transfer, fromId, toId, operatorId));
         }
